VoiceMap/requirements.txt
# LiveKit Agent Dependencies for VoiceMap
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0  
livekit-plugins-silero>=0.6.0
websockets>=12.0
numpy>=1.24.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
PyJWT>=2.8.0